+++ sdk/cognitivelanguage/azure-ai-textanalytics/samples/sample_recognize_pii_entities_with_confidence_score.py
@@ -57,8 +57,8 @@ def sample_pii_with_confidence_score():
     threshold = ConfidenceScoreThreshold(
         default=0.3,
         overrides=[
-            ConfidenceScoreThresholdOverride(value=0.9, entity="USSocialSecurityNumber"),
-            ConfidenceScoreThresholdOverride(value=0.9, entity="Email"),
+            ConfidenceScoreThresholdOverride(value=0.9, entity="USSocialSecurityNumber", language="en"),
+            ConfidenceScoreThresholdOverride(value=0.9, entity="Email", language="en"),
         ],
     )