+++ sdk/cognitivelanguage/azure-ai-textanalytics/tests/test_text_sentiment_async.py
@@ -10,10 +10,9 @@ import pytest
 
 from devtools_testutils import (
     AzureRecordedTestCase,
-    EnvironmentVariableLoader,
+    PowerShellPreparer,
 )
 from devtools_testutils.aio import recorded_by_proxy_async
-from azure.core.credentials import AzureKeyCredential
 from azure.ai.textanalytics.aio import TextAnalysisClient
 from azure.ai.textanalytics.models import (
     MultiLanguageTextInput,
@@ -25,24 +24,28 @@ from azure.ai.textanalytics.models import (
 )
 
 TextAnalysisPreparer = functools.partial(
-    EnvironmentVariableLoader,
+    PowerShellPreparer,
     "text_analysis",
     text_analysis_endpoint="https://Sanitized.cognitiveservices.azure.com/",
-    text_analysis_key="fake_key",
 )
 
 
 class TestTextAnalysisAsync(AzureRecordedTestCase):
-    def create_client(self, endpoint: str, key: str) -> TextAnalysisClient:
-        return TextAnalysisClient(endpoint, AzureKeyCredential(key))
+    def create_client(self, endpoint: str) -> TextAnalysisClient:
+        credential = self.get_credential(TextAnalysisClient, is_async=True)
+        return self.create_client_from_credential(
+            TextAnalysisClient,
+            credential=credential,
+            endpoint=endpoint,
+        )
 
 
 class TestTextAnalysisCaseAsync(TestTextAnalysisAsync):
     @TextAnalysisPreparer()
     @recorded_by_proxy_async
     @pytest.mark.asyncio
-    async def test_text_sentiment_async(self, text_analysis_endpoint, text_analysis_key):
-        async with self.create_client(text_analysis_endpoint, text_analysis_key) as client:
+    async def test_text_sentiment_async(self, text_analysis_endpoint):
+        async with self.create_client(text_analysis_endpoint) as client:
             text_a = (
                 "The food and service were unacceptable, but the concierge were nice. After talking to them about the "
                 "quality of the food and the process to get room service they refunded the money we spent at the "